\name{exportToSnnsNetFile}
\alias{exportToSnnsNetFile}
\title{Export the net to a file in the original SNNS file format}
\usage{exportToSnnsNetFile(object, filename, netname="RSNNS_untitled")
}
\description{Export the net that is present in the \code{\link{rsnns}} object in the 
original (.net) SNNS file format.}
\arguments{\item{object}{the \code{\link{rsnns}} object}
\item{filename}{path and filename to be written to}
\item{netname}{name that is given to the network in the file}
}

