\name{getNormParameters}
\alias{getNormParameters}
\title{Get normalization parameters of the input data}
\usage{getNormParameters(x)
}
\description{Get the normalization parameters that are appended by \code{\link{normalizeData}}
as attributes to the input data.}
\details{This function is equivalent to calling \code{attr(x, "normParams")}.}
\value{the parameters generated by an earlier call to \code{\link{normalizeData}}}
\seealso{\code{\link{normalizeData}}, \code{\link{denormalizeData}}}
\arguments{\item{x}{input data}
}

