% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inputColumns}
\alias{inputColumns}
\title{Get the columns that are inputs}
\usage{
inputColumns(patterns)
}
\arguments{
\item{patterns}{matrix or data.frame containing the patterns}
}
\description{
This function extracts all columns from a matrix whose column names begin with "in".
The example data of this package follows this naming convention.
}

