\name{outputColumns}
\alias{outputColumns}
\title{Function to get the columns that are targets.}
\usage{outputColumns(patterns)
}
\description{Function to get the columns that are targets.}
\details{This function extracts all columns from a matrix whose column names begin with "out".
The example data of this package follows this naming convention.}
\arguments{\item{patterns}{matrix or data.frame containing the patterns}
}

