% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_class.R
\name{plotRegressionError}
\alias{plotRegressionError}
\title{Plot a regression error plot}
\usage{
plotRegressionError(targets, fits, ...)
}
\arguments{
\item{targets}{the target values}

\item{fits}{the values predicted/fitted by the model}

\item{...}{parameters passed to \code{plot}}
}
\description{
The plot shows target values on the x-axis and fitted/predicted values on the y-axis. 
The optimal fit would yield a line through zero with gradient one.
This optimal line is shown in black color. A linear fit to the actual data
is shown in red color.
}
