% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.rsnns}
\alias{predict.rsnns}
\title{Generic predict function for rsnns object}
\usage{
\method{predict}{rsnns}(object, newdata, ...)
}
\arguments{
\item{object}{the \code{\link{rsnns}} object}

\item{newdata}{the new input data which is used for prediction}

\item{...}{additional function parameters (currently not used)}
}
\value{
the predicted values
}
\description{
Predict values using the given network.
}

