\name{readResFile}
\alias{readResFile}
\title{Rudimentary parser for res files.}
\usage{readResFile(filename)
}
\description{Rudimentary parser for res files.}
\details{This function contains a rudimentary parser for SNNS .res files. It
is completely implemented in R and doesn't make use of SNNS functionality.}
\value{a matrix containing the predicted values that were found in the .res file}
\arguments{\item{filename}{the name of the .res file}
}

