% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{savePatFile}
\alias{savePatFile}
\title{Save data to a pat file}
\usage{
savePatFile(inputs, targets, filename)
}
\arguments{
\item{inputs}{a matrix with input values}

\item{targets}{a matrix with target values}

\item{filename}{the name of the .pat file}
}
\description{
This function generates an \link{SnnsR-class} object, loads the given data there 
as a pattern set and then uses the functionality of SNNS to save the data as a .pat file.
}
