\name{snnsData}
\alias{snnsData}
\alias{snnsData}
\title{Example data of the package.}
\description{Example data of the package.}
\details{This is data from the original SNNS examples directory ported to R and stored as one list.
The function \code{\link{readPatFile}} was used to parse all pattern files (.pat) from the 
original SNNS examples directory. Due to limitations of that function, pattern files
containing patterns with variable size were omitted.}
\docType{data}
\keyword{data}
\examples{data(snnsData)
names(snnsData)}

