\name{SnnsR-class}
\alias{SnnsR-class}
\alias{SnnsR}
\title{The main class of the package.}
\description{The main class of the package. Each instance contains a pointer to 
a C++ object of type SnnsCLib, i.e. an instance of the SNNS kernel.}
\section{Slots}{\describe{\item{\code{snnsCLibPointer}:}{}}}

