\name{SnnsRObject$extractNetInfo}
\alias{SnnsR__extractNetInfo}
\title{Get characteristics of the network.}
\usage{\S4method{extractNetInfo}{SnnsR}()}
\description{Get characteristics of the network.}
\details{The returned list has three members: 
\itemize{
\item infoHeader general information about the network
\item unitDefinitions information about the units
\item fullWeightMatrix weight matrix of the connections
}}
\value{a list of data frames containing information extracted from the network.}
\alias{extractNetInfo,SnnsR-method}
\alias{SnnsRObject$extractNetInfo}

