\name{SnnsRObject$extractPatterns}
\alias{SnnsR__extractPatterns}
\title{Extract the current pattern set to a matrix.}
\usage{\S4method{extractPatterns}{SnnsR}()}
\description{Extract the current pattern set to a matrix.}
\details{Extracts all patterns of the current pattern set and
returns them as a matrix. Columns are named with the prefix "in" or "out", respectively.}
\value{a matrix containing the patterns of the currently loaded patern set.}
\alias{extractPatterns,SnnsR-method}
\alias{SnnsRObject$extractPatterns}

