\name{SnnsRObject$genericPredictCurrPatSet}
\alias{SnnsR__genericPredictCurrPatSet}
\title{Predict values with a trained net.}
\usage{\S4method{genericPredictCurrPatSet}{SnnsR}(units, updateFuncParams=c(0.0))}
\description{Predict values with a trained net.}
\details{Generic function for prediction with a trained net.}
\value{the predicted values}
\alias{genericPredictCurrPatSet,SnnsR-method}
\alias{SnnsRObject$genericPredictCurrPatSet}
\arguments{\item{units}{the units that define the output}
\item{updateFuncParams}{the parameters for the update function (the function has to be already set)}
}

