\name{SnnsRObject$getAllOutputUnits}
\alias{SnnsR__getAllOutputUnits}
\title{Get all output units of the net.}
\usage{\S4method{getAllOutputUnits}{SnnsR}()}
\description{Get all output units of the net.}
\value{a vector with integer numbers identifying the units.}
\alias{getAllOutputUnits,SnnsR-method}
\alias{SnnsRObject$getAllOutputUnits}
\seealso{\link{SnnsRObject$getAllUnitsTType}}

