\name{SnnsRObject$getAllUnits}
\alias{SnnsR__getAllUnits}
\title{Get all units present in the net.}
\usage{\S4method{getAllUnits}{SnnsR}()}
\description{Get all units present in the net.}
\value{a vector with integer numbers identifying the units.}
\alias{getAllUnits,SnnsR-method}
\alias{SnnsRObject$getAllUnits}

