\name{SnnsRObject$getCompleteWeightMatrix}
\alias{SnnsR__getCompleteWeightMatrix}
\title{Get the complete weight matrix.}
\usage{\S4method{getCompleteWeightMatrix}{SnnsR}(setDimNames)}
\description{Get the complete weight matrix.}
\details{Get a weight matrix containing all weights of all neurons present in the net.}
\value{the complete weight matrix}
\alias{getCompleteWeightMatrix,SnnsR-method}
\alias{SnnsRObject$getCompleteWeightMatrix}
\arguments{\item{setDimNames}{indicates, whether names of units are extracted and set as row/col names in the weight matrix}
}

