\name{SnnsRObject$getInfoHeader}
\alias{SnnsR__getInfoHeader}
\title{Get an info header of the network.}
\usage{\S4method{getInfoHeader}{SnnsR}()}
\description{Get an info header of the network.}
\value{a data frame containing some general characteristics of the network.}
\alias{getInfoHeader,SnnsR-method}
\alias{SnnsRObject$getInfoHeader}

