\name{SnnsRObject$getSiteDefinitions}
\alias{SnnsR__getSiteDefinitions}
\title{Get the sites definitions of the network.}
\usage{\S4method{getSiteDefinitions}{SnnsR}()}
\description{Get the sites definitions of the network.}
\value{a data frame containing information about all sites present in the network.}
\alias{getSiteDefinitions,SnnsR-method}
\alias{SnnsRObject$getSiteDefinitions}

