\name{SnnsRObject$getTypeDefinitions}
\alias{SnnsR__getTypeDefinitions}
\title{Get the FType definitions of the network.}
\usage{\S4method{getTypeDefinitions}{SnnsR}()}
\description{Get the FType definitions of the network.}
\value{a data frame containing information about FType units present in the network.}
\alias{getTypeDefinitions,SnnsR-method}
\alias{SnnsRObject$getTypeDefinitions}

