\name{SnnsRObject$getUnitDefinitions}
\alias{SnnsR__getUnitDefinitions}
\title{Get the unit definitions of the network.}
\usage{\S4method{getUnitDefinitions}{SnnsR}()}
\description{Get the unit definitions of the network.}
\value{a data frame containing information about all units present in the network.}
\alias{getUnitDefinitions,SnnsR-method}
\alias{SnnsRObject$getUnitDefinitions}

