\name{SnnsRObject$getUnitsByName}
\alias{SnnsR__getUnitsByName}
\title{Find all units whose name begins with a given prefix.}
\usage{\S4method{getUnitsByName}{SnnsR}(prefix)}
\description{Find all units whose name begins with a given prefix.}
\value{a vector with integer numbers identifying the units.}
\alias{getUnitsByName,SnnsR-method}
\alias{SnnsRObject$getUnitsByName}
\arguments{\item{prefix}{a prefix that the names of the units to find have.}
}

