\name{SnnsRObject$resetRSNNS}
\alias{SnnsR__resetRSNNS}
\title{Reset the SnnsR object.}
\usage{\S4method{resetRSNNS}{SnnsR}()}
\description{SnnsR low-level function to delete all pattern sets and 
delete the current net in the \code{\link{SnnsR-class}} object.}
\alias{resetRSNNS,SnnsR-method}
\alias{SnnsRObject$resetRSNNS}

