\name{SnnsRObject$somPredictComponentMaps}
\alias{SnnsR__somPredictComponentMaps}
\title{Calculate the som component maps.}
\usage{\S4method{somPredictComponentMaps}{SnnsR}(updateFuncParams=c(0.0, 0.0, 1.0))}
\description{Calculate the som component maps.}
\value{a matrix containing all componant maps as 1d vectors}
\alias{somPredictComponentMaps,SnnsR-method}
\alias{SnnsRObject$somPredictComponentMaps}
\seealso{\code{\link{som}}}
\arguments{\item{updateFuncParams}{parameters passed to the networks update function}
}

