\name{SnnsRObject$somPredictCurrPatSetWinnersSpanTree}
\alias{SnnsR__somPredictCurrPatSetWinnersSpanTree}
\title{Get the spanning tree of the som, calculated directly by SNNS.}
\usage{\S4method{somPredictCurrPatSetWinnersSpanTree}{SnnsR}()}
\description{Get the spanning tree of the som, calculated directly by SNNS.}
\value{the spanning tree, which is the som, showing for each unit a number identifying 
the last pattern for which this unit won. (We note that, also if there are more than 
one patterns, only the last one is saved)}
\alias{somPredictCurrPatSetWinnersSpanTree,SnnsR-method}
\alias{SnnsRObject$somPredictCurrPatSetWinnersSpanTree}
\seealso{\code{\link{som}}}

