% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SnnsRObject$createPatSet}
\alias{SnnsRObject$createPatSet}
\alias{SnnsR__createPatSet}
\alias{createPatSet,SnnsR-method}
\title{Create a pattern set}
\usage{
\S4method{createPatSet}{SnnsR}(inputs, targets)
}
\arguments{
\item{inputs}{the input values}

\item{targets}{the target values}
}
\value{
a list with elements \code{err} and \code{set_no}. The latter one identifies the pattern set within the \code{\link{SnnsR-class}} object
}
\description{
SnnsR low-level function to create a pattern set in the SNNS kernel from
the values given, so that they are available in the SNNS kernel for use.
}

