% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SnnsRObject$getAllUnits}
\alias{SnnsRObject$getAllUnits}
\alias{SnnsR__getAllUnits}
\alias{getAllUnits,SnnsR-method}
\title{Get all units present in the net.}
\usage{
\S4method{getAllUnits}{SnnsR}()
}
\value{
a vector with integer numbers identifying the units.
}
\description{
Get all units present in the net.
}

