% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_extractNetInfo.R
\name{SnnsRObject$getAllUnitsTType}
\alias{SnnsRObject$getAllUnitsTType}
\alias{SnnsR__getAllUnitsTType}
\alias{getAllUnitsTType,SnnsR-method}
\title{Get all units in the net of a certain \code{ttype}.}
\usage{
\S4method{getAllUnitsTType}{SnnsR}(ttype)
}
\arguments{
\item{ttype}{a string containing the \code{ttype}.}
}
\value{
a vector with integer numbers identifying the units.
}
\description{
SnnsR low-level function to get all units in the net of a certain \code{ttype}. 
Possible \code{ttype} defined by SNNS are, among others:
"UNIT_OUTPUT", "UNIT_INPUT", and "UNIT_HIDDEN". For a full list, 
call \code{RSNNS:::SnnsDefines$topologicalUnitTypes}
As this is an SnnsR low-level function, you may want to have a look 
at \code{\link{SnnsR-class}} to find out how to properly use it.
}
\seealso{
\link{SnnsRObject$getAllOutputUnits}, \link{SnnsRObject$getAllInputUnits}, \link{SnnsRObject$getAllHiddenUnits}
}
