\name{SnnsRObject$getInfoHeader}
\alias{getInfoHeader,SnnsR-method}
\alias{SnnsR__getInfoHeader}
\alias{SnnsRObject$getInfoHeader}
\title{Get an info header of the network.}
\usage{
  \S4method{getInfoHeader}{SnnsR}()
}
\value{
  a data frame containing some general characteristics of
  the network.
}
\description{
  Get an info header of the network.
}

