\name{SnnsRObject$getSiteDefinitions}
\alias{getSiteDefinitions,SnnsR-method}
\alias{SnnsR__getSiteDefinitions}
\alias{SnnsRObject$getSiteDefinitions}
\title{Get the sites definitions of the network.}
\usage{
  \S4method{getSiteDefinitions}{SnnsR}()
}
\value{
  a data frame containing information about all sites
  present in the network.
}
\description{
  Get the sites definitions of the network.
}

