\name{SnnsRObject$getTypeDefinitions}
\alias{getTypeDefinitions,SnnsR-method}
\alias{SnnsR__getTypeDefinitions}
\alias{SnnsRObject$getTypeDefinitions}
\title{Get the FType definitions of the network.}
\usage{
  \S4method{getTypeDefinitions}{SnnsR}()
}
\value{
  a data frame containing information about FType units
  present in the network.
}
\description{
  Get the FType definitions of the network.
}

