\name{SnnsRObject$getUnitDefinitions}
\alias{getUnitDefinitions,SnnsR-method}
\alias{SnnsR__getUnitDefinitions}
\alias{SnnsRObject$getUnitDefinitions}
\title{Get the unit definitions of the network.}
\usage{
  \S4method{getUnitDefinitions}{SnnsR}()
}
\value{
  a data frame containing information about all units
  present in the network.
}
\description{
  Get the unit definitions of the network.
}

