\name{SnnsRObject$getUnitsByName}
\alias{getUnitsByName,SnnsR-method}
\alias{SnnsR__getUnitsByName}
\alias{SnnsRObject$getUnitsByName}
\title{Find all units whose name begins with a given prefix.}
\usage{
  \S4method{getUnitsByName}{SnnsR}(prefix)
}
\arguments{
  \item{prefix}{a prefix that the names of the units to
  find have.}
}
\value{
  a vector with integer numbers identifying the units.
}
\description{
  Find all units whose name begins with a given prefix.
}

