\name{SnnsRObject$resetRSNNS}
\alias{resetRSNNS,SnnsR-method}
\alias{SnnsR__resetRSNNS}
\alias{SnnsRObject$resetRSNNS}
\title{Reset the SnnsR object.}
\usage{
  \S4method{resetRSNNS}{SnnsR}()
}
\description{
  SnnsR low-level function to delete all pattern sets and
  delete the current net in the \code{\link{SnnsR-class}}
  object.
}

