% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_patterns.R
\name{SnnsRObject$somPredictComponentMaps}
\alias{SnnsRObject$somPredictComponentMaps}
\alias{SnnsR__somPredictComponentMaps}
\alias{somPredictComponentMaps,SnnsR-method}
\title{Calculate the som component maps}
\usage{
\S4method{somPredictComponentMaps}{SnnsR}(updateFuncParams=c(0.0, 0.0, 1.0))
}
\arguments{
\item{updateFuncParams}{parameters passed to the networks update function}
}
\value{
a matrix containing all componant maps as 1d vectors
}
\description{
SnnsR low-level function to calculate the som component maps.
}
\seealso{
\code{\link{som}}
}
