% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_patterns.R
\name{SnnsRObject$whereAreResults}
\alias{SnnsRObject$whereAreResults}
\alias{SnnsR__whereAreResults}
\alias{whereAreResults,SnnsR-method}
\title{Get a list of output units of a net}
\usage{
\S4method{whereAreResults}{SnnsR}(outputMethod="output")
}
\arguments{
\item{outputMethod}{a string defining the output method of the net. Possible values are: "art1", "art2", "artmap", "assoz", "som", "output".}
}
\value{
a list of numbers identifying the units
}
\description{
SnnsR low-level function to get a list of output units of a net.
}
\details{
Depending on the network architecture, output is present in hidden units, in
output units, etc. In some network types, the output units have a certain
name prefix in SNNS. This function finds the output units according to
certain network types. The type is specified by \code{outputMethod}. If the
given \code{outputMethod} is unknown, the function defaults to "output".
}
