% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_class.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Computes a confusion matrix}
\usage{
confusionMatrix(targets, predictions)
}
\arguments{
\item{targets}{the known, correct target values}

\item{predictions}{the corresponding predictions of a method for the targets}
}
\value{
the confusion matrix
}
\description{
The confusion matrix shows how many times a pattern
with the real class x was classified as class y. A perfect method
should result in a diagonal matrix. All values not on the diagonal
are errors of the method.
}
\details{
If the class labels are not already encoded, they are encoded using \code{\link{encodeClassLabels}} 
(with default values).
}
