% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsnns.R
\name{exportToSnnsNetFile}
\alias{exportToSnnsNetFile}
\title{Export the net to a file in the original SNNS file format}
\usage{
exportToSnnsNetFile(object, filename, netname = "RSNNS_untitled")
}
\arguments{
\item{object}{the \code{\link{rsnns}} object}

\item{filename}{path and filename to be written to}

\item{netname}{name that is given to the network in the file}
}
\description{
Export the net that is present in the \code{\link{rsnns}} object in the 
original (.net) SNNS file format.
}
