% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeData.R
\name{getNormParameters}
\alias{getNormParameters}
\title{Get normalization parameters of the input data}
\usage{
getNormParameters(x)
}
\arguments{
\item{x}{input data}
}
\value{
the parameters generated by an earlier call to \code{\link{normalizeData}}
}
\description{
Get the normalization parameters that are appended by \code{\link{normalizeData}}
as attributes to the input data.
}
\details{
This function is equivalent to calling \code{attr(x, "normParams")}.
}
\seealso{
\code{\link{normalizeData}}, \code{\link{denormalizeData}}
}
