% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getSnnsRFunctionTable}
\alias{getSnnsRFunctionTable}
\title{Get SnnsR function table}
\usage{
getSnnsRFunctionTable()
}
\value{
a data.frame with columns:
\item{name}{name of the function}
\item{type}{the type of the function (learning, init, update,...)}
\item{#inParams}{the number of input parameters of the function}
\item{#outParams}{the number of output parameters of the function}
}
\description{
Get the function table of available SNNS functions.
}
