\name{inputColumns}
\alias{inputColumns}
\title{Function to get the columns that are inputs.}
\usage{inputColumns(patterns)
}
\description{Function to get the columns that are inputs.}
\details{This function extracts all columns from a matrix whose column names begin with "in".
The example data of this package follows this naming convention.}
\arguments{\item{patterns}{matrix or data.frame containing the patterns}
}

