% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{outputColumns}
\alias{outputColumns}
\title{Get the columns that are targets}
\usage{
outputColumns(patterns)
}
\arguments{
\item{patterns}{matrix or data.frame containing the patterns}
}
\description{
This function extracts all columns from a matrix whose column names begin with "out".
The example data of this package follows this naming convention.
}
