% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotActMap}
\alias{plotActMap}
\title{Plot activation map}
\usage{
plotActMap(x, ...)
}
\arguments{
\item{x}{the input data matrix}

\item{...}{parameters passed to \code{image}}
}
\description{
Plot an activation map as a heatmap.
}
\seealso{
\link{vectorToActMap} \link{matrixToActMapList}
}

