% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_class.R
\name{plotIterativeError}
\alias{plotIterativeError}
\alias{plotIterativeError.rsnns}
\title{Plot iterative errors of an rsnns object}
\usage{
plotIterativeError(object, ...)

\method{plotIterativeError}{rsnns}(object, ...)
}
\arguments{
\item{object}{a rsnns object}

\item{...}{parameters passed to \code{plot}}
}
\description{
Plot the iterative training and test error of the net of this \code{\link{rsnns}} object.
}
\details{
Plots (if present) the class members \code{IterativeFitError} (as black line) and 
\code{IterativeTestError} (as red line).
}
