\name{plotIterativeError}
\alias{plotIterativeError}
\title{Plot the iterative training and test error of the net of this rsnns object.}
\usage{plotIterativeError(object, ...)
\method{plotIterativeError}{reg_class}(object, ...)
}
\description{Plot the iterative training and test error of the net of this rsnns object.}
\details{\code{plotIterativeError}: Plots (if present) the class members \code{IterativeFitError} (as black line) and 
\code{IterativeTestError} (as red line).

}
\arguments{\item{object}{the object to which to apply plotIterativeError}
\item{...}{additional function parameters}
}
\alias{plotIterativeError.reg_class}

