\name{plotROC}
\alias{plotROC}
\title{Plot a ROC curve}
\usage{
  plotROC(T, D, ...)
}
\arguments{
  \item{T}{predictions}

  \item{D}{targets}

  \item{...}{parameters passed to plot}
}
\description{
  This function plots a receiver operating characteristic
  (ROC) curve.
}
\author{
  Code is taken from R news Volume 4/1, June 2004.
}
\references{
  R news Volume 4/1, June 2004
}

