\name{predict.rsnns}
\alias{predict.rsnns}
\title{Generic predict function for rsnns object}
\usage{\method{predict}{rsnns}(object, newdata, ...)
}
\description{Predict values using the given network.}
\value{the predicted values}
\arguments{\item{object}{the \code{\link{rsnns}} object}
\item{newdata}{the new input data which is used for prediction}
\item{...}{additional function parameters (currently not used)}
}

