% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsnns.R
\name{print.rsnns}
\alias{print.rsnns}
\title{Generic print function for rsnns objects}
\usage{
\method{print}{rsnns}(x, ...)
}
\arguments{
\item{x}{the \code{\link{rsnns}} object}

\item{...}{additional function parameters (currently not used)}
}
\description{
Print out some characteristics of an \code{\link{rsnns}} object.
}
