\name{print.rsnns}
\alias{print.rsnns}
\title{Generic print function for rsnns objects.}
\usage{\method{print}{rsnns}(x, ...)
}
\description{Generic print function for \code{rsnns} objects.}
\details{Print out some characteristics of an \code{\link{rsnns}} object.}
\arguments{\item{x}{the \code{\link{rsnns}} object}
\item{...}{additional function parameters (currently not used)}
}

