% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readResFile}
\alias{readResFile}
\title{Rudimentary parser for res files.}
\usage{
readResFile(filename)
}
\arguments{
\item{filename}{the name of the .res file}
}
\value{
a matrix containing the predicted values that were found in the .res file
}
\description{
This function contains a rudimentary parser for SNNS .res files. It
is completely implemented in R and doesn't make use of SNNS functionality.
}

