% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{resolveSnnsRDefine}
\alias{resolveSnnsRDefine}
\title{Resolve a define of the SNNS kernel}
\usage{
resolveSnnsRDefine(defList, def)
}
\arguments{
\item{defList}{the defines-list from which to resolve the define from}

\item{def}{the name of the define}
}
\value{
the value of the define
}
\description{
Resolve a define of the SNNS kernel using a defines-list.
All defines-lists present can be shown with \code{RSNNS:::SnnsDefines}.
}
\examples{
resolveSnnsRDefine("topologicalUnitTypes","UNIT_HIDDEN")
}
\seealso{
\code{\link{getSnnsRDefine}}
}
