% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{snnsData}
\alias{snnsData}
\title{Example data of the package}
\description{
This is data from the original SNNS examples directory ported to R and stored as one list.
The function \code{\link{readPatFile}} was used to parse all pattern files (.pat) from the 
original SNNS examples directory. Due to limitations of that function, pattern files
containing patterns with variable size were omitted.
}
\examples{
data(snnsData)
names(snnsData)
}
\keyword{data}
