\name{SnnsRObject$getAllHiddenUnits}
\alias{SnnsR__getAllHiddenUnits}
\title{Get all hidden units of the net.}
\usage{\S4method{getAllHiddenUnits}{SnnsR}()}
\description{Get all hidden units of the net.}
\value{a vector with integer numbers identifying the units.}
\alias{getAllHiddenUnits,SnnsR-method}
\alias{SnnsRObject$getAllHiddenUnits}
\seealso{\link{SnnsRObject$getAllUnitsTType}}

