\name{SnnsRObject$getAllInputUnits}
\alias{SnnsR__getAllInputUnits}
\title{Get all input units of the net.}
\usage{\S4method{getAllInputUnits}{SnnsR}()}
\description{Get all input units of the net.}
\value{a vector with integer numbers identifying the units.}
\alias{getAllInputUnits,SnnsR-method}
\alias{SnnsRObject$getAllInputUnits}
\seealso{\link{SnnsRObject$getAllUnitsTType}}

