\name{SnnsRObject$resetRSNNS}
\alias{SnnsR__resetRSNNS}
\title{Reset the SnnsR-class object.}
\usage{\S4method{resetRSNNS}{SnnsR}()}
\description{Reset the \code{SnnsR-class} object.}
\details{Delete all pattern sets and delete the current net in the \code{\link{SnnsR-class}} object.}
\alias{resetRSNNS,SnnsR-method}
\alias{SnnsRObject$resetRSNNS}

