% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_extractNetInfo.R
\name{SnnsRObject$extractNetInfo}
\alias{SnnsRObject$extractNetInfo}
\alias{SnnsR__extractNetInfo}
\alias{extractNetInfo,SnnsR-method}
\title{Get characteristics of the network.}
\usage{
\S4method{extractNetInfo}{SnnsR}()
}
\value{
a list of data frames containing information extracted from the network.
}
\description{
The returned list has three members: 
\itemize{
\item infoHeader general information about the network
\item unitDefinitions information about the units
\item fullWeightMatrix weight matrix of the connections
}
}
