% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_extractNetInfo.R
\name{SnnsRObject$getAllHiddenUnits}
\alias{SnnsRObject$getAllHiddenUnits}
\alias{SnnsR__getAllHiddenUnits}
\alias{getAllHiddenUnits,SnnsR-method}
\title{Get all hidden units of the net}
\usage{
\S4method{getAllHiddenUnits}{SnnsR}()
}
\value{
a vector with integer numbers identifying the units.
}
\description{
SnnsR low-level function to get all units from the net with the ttype "UNIT_HIDDEN".
This function calls \code{\link{SnnsRObject$getAllUnitsTType}} with the parameter "UNIT_HIDDEN".
}
\seealso{
\link{SnnsRObject$getAllUnitsTType}
}
