% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_extractNetInfo.R
\name{SnnsRObject$getAllInputUnits}
\alias{SnnsRObject$getAllInputUnits}
\alias{SnnsR__getAllInputUnits}
\alias{getAllInputUnits,SnnsR-method}
\title{Get all input units of the net}
\usage{
\S4method{getAllInputUnits}{SnnsR}()
}
\value{
a vector with integer numbers identifying the units.
}
\description{
SnnsR low-level function to get all units from the net with the ttype "UNIT_INPUT".
This function calls \code{\link{SnnsRObject$getAllUnitsTType}} with the parameter "UNIT_INPUT".
}
\seealso{
\link{SnnsRObject$getAllUnitsTType}
}
