% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_extractNetInfo.R
\name{SnnsRObject$getAllOutputUnits}
\alias{SnnsRObject$getAllOutputUnits}
\alias{SnnsR__getAllOutputUnits}
\alias{getAllOutputUnits,SnnsR-method}
\title{Get all output units of the net.}
\usage{
\S4method{getAllOutputUnits}{SnnsR}()
}
\value{
a vector with integer numbers identifying the units.
}
\description{
SnnsR low-level function to get all units from the net with the ttype "UNIT_OUTPUT".
This function calls \code{\link{SnnsRObject$getAllUnitsTType}} with the parameter "UNIT_OUTPUT".
}
\seealso{
\code{\link{SnnsRObject$getAllUnitsTType}}
}
