\name{SnnsRObject$getAllUnits}
\alias{getAllUnits,SnnsR-method}
\alias{SnnsR__getAllUnits}
\alias{SnnsRObject$getAllUnits}
\title{Get all units present in the net.}
\usage{
  \S4method{getAllUnits}{SnnsR}()
}
\value{
  a vector with integer numbers identifying the units.
}
\description{
  Get all units present in the net.
}

