% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_extractNetInfo.R
\name{SnnsRObject$getCompleteWeightMatrix}
\alias{SnnsRObject$getCompleteWeightMatrix}
\alias{SnnsR__getCompleteWeightMatrix}
\alias{getCompleteWeightMatrix,SnnsR-method}
\title{Get the complete weight matrix.}
\usage{
\S4method{getCompleteWeightMatrix}{SnnsR}(setDimNames)
}
\arguments{
\item{setDimNames}{indicates, whether names of units are extracted and set as row/col names in the weight matrix}
}
\value{
the complete weight matrix
}
\description{
Get a weight matrix containing all weights of all neurons present in the net.
}
