% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_extractNetInfo.R
\name{SnnsRObject$getInfoHeader}
\alias{SnnsRObject$getInfoHeader}
\alias{SnnsR__getInfoHeader}
\alias{getInfoHeader,SnnsR-method}
\title{Get an info header of the network.}
\usage{
\S4method{getInfoHeader}{SnnsR}()
}
\value{
a data frame containing some general characteristics of the network.
}
\description{
Get an info header of the network.
}
