% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_extractNetInfo.R
\name{SnnsRObject$getSiteDefinitions}
\alias{SnnsRObject$getSiteDefinitions}
\alias{SnnsR__getSiteDefinitions}
\alias{getSiteDefinitions,SnnsR-method}
\title{Get the sites definitions of the network.}
\usage{
\S4method{getSiteDefinitions}{SnnsR}()
}
\value{
a data frame containing information about all sites present in the network.
}
\description{
Get the sites definitions of the network.
}
