% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_extractNetInfo.R
\name{SnnsRObject$getTypeDefinitions}
\alias{SnnsRObject$getTypeDefinitions}
\alias{SnnsR__getTypeDefinitions}
\alias{getTypeDefinitions,SnnsR-method}
\title{Get the FType definitions of the network.}
\usage{
\S4method{getTypeDefinitions}{SnnsR}()
}
\value{
a data frame containing information about FType units present in the network.
}
\description{
Get the FType definitions of the network.
}
