% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SnnsRObject$getUnitDefinitions}
\alias{SnnsRObject$getUnitDefinitions}
\alias{SnnsR__getUnitDefinitions}
\alias{getUnitDefinitions,SnnsR-method}
\title{Get the unit definitions of the network.}
\usage{
\S4method{getUnitDefinitions}{SnnsR}()
}
\value{
a data frame containing information about all units present in the network.
}
\description{
Get the unit definitions of the network.
}

