% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_extractNetInfo.R
\name{SnnsRObject$getUnitsByName}
\alias{SnnsRObject$getUnitsByName}
\alias{SnnsR__getUnitsByName}
\alias{getUnitsByName,SnnsR-method}
\title{Find all units whose name begins with a given prefix.}
\usage{
\S4method{getUnitsByName}{SnnsR}(prefix)
}
\arguments{
\item{prefix}{a prefix that the names of the units to find have.}
}
\value{
a vector with integer numbers identifying the units.
}
\description{
Find all units whose name begins with a given prefix.
}
