% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnnsR_util.R
\name{SnnsRObject$resetRSNNS}
\alias{SnnsRObject$resetRSNNS}
\alias{SnnsR__resetRSNNS}
\alias{resetRSNNS,SnnsR-method}
\title{Reset the SnnsR object.}
\usage{
\S4method{resetRSNNS}{SnnsR}()
}
\description{
SnnsR low-level function to delete all pattern sets and 
delete the current net in the \code{\link{SnnsR-class}} object.
}
