% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsnns.R
\name{weightMatrix}
\alias{weightMatrix}
\alias{weightMatrix.rsnns}
\title{Function to extract the weight matrix of an rsnns object}
\usage{
weightMatrix(object, ...)

\method{weightMatrix}{rsnns}(object, ...)
}
\arguments{
\item{object}{the \code{\link{rsnns}} object}

\item{...}{additional function parameters (currently not used)}
}
\value{
a matrix with all weights from all neurons present in the net.
}
\description{
The function calls \code{\link{SnnsRObject$getCompleteWeightMatrix}} and returns its result.
}
